requirements.txt
flet
PyYAML
Jinja2
GitPython
requests
dash
plotly